+++ include/H2DE/H2DE_settings.h
@@ -1,0 +1,11 @@
+// #ifndef H2DE_SETTINGS_H
+// #define H2DE_SETTINGS_H
+
+// #include <inih/INIReader.h>
+
+// class H2DE_Settings {
+// public:
+    
+// };
+
+// #endif